@@ -3,29 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.11.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-regular-svg-icons": "^5.11.2",
-    "@fortawesome/free-solid-svg-icons": "^5.11.2",
-    "@fortawesome/react-fontawesome": "^0.1.7",
-    "bootstrap3": "^3.3.5",
-    "query-string": "^6.8.3",
-    "react": "^16.12.0",
-    "react-bootstrap": "^0.32.4",
-    "react-cookie": "^4.0.1",
-    "react-dom": "^16.12.0",
-    "react-keydown": "^1.9.12",
-    "react-redux": "^7.1.3",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
-    "redux": "^4.0.4",
+    "@fortawesome/fontawesome-free": "^6.4.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-regular-svg-icons": "^6.3.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "bootstrap": "^5.2.3",
+    "clarin-bootstrap": "file:../clarin-bootstrap",
+    "core-js": "^3.30.1",
+    "query-string": "^8.1.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.4",
+    "react-cookie": "^4.1.1",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router": "^6.10.0",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
-    "webpack": "4.41.2",
-    "core-js": "^3.6.0",
-    "typescript": "^3.7.3",
-    "whatwg-fetch": "^3.0.0"
+    "redux-thunk": "^2.4.2",
+    "webpack": "5.79.0",
+    "whatwg-fetch": "^3.6.2"
   },
   "scripts": {
     "start": "react-scripts start --host localhost --port 3000",
